# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project manager API",
        "version": "0.1"
    },
    "paths": {
...
        "/situation": {
            "post": {
                "tags": [
                    "Situations"
                ],
                "summary": "Create situation",
                "description": "Create a situation.",
                "operationId": "1d31dd02175cde51a952d4222e92d9c2",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Authorization token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer 1|y6VIcnEFVMpQvOHStKU0dcXD8CKEdJ0nQuW3anF6"
                        }
                    },
                    {
                        "name": "project_id",
                        "in": "query",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Situation name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Situation description",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Situation status",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/situation/{id}": {
            "get": {
                "tags": [
                    "Situations"
                ],
                "summary": "Get situation details",
                "description": "Get the situation.",
                "operationId": "c33e71c66995439c70f148da78d25ecb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Situation id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Authorization token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer 1|y6VIcnEFVMpQvOHStKU0dcXD8CKEdJ0nQuW3anF6"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Situation"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Situations"
                ],
                "summary": "Delete situation",
                "description": "Delete situation.",
                "operationId": "19bf29919ad1c266fea7c6b13eb3357c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Situation id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Authorization token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer 1|y6VIcnEFVMpQvOHStKU0dcXD8CKEdJ0nQuW3anF6"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Situations"
                ],
                "summary": "Update situation details",
                "description": "Update situation.",
                "operationId": "e4ecb46703ca4f71961a740fa9731891",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Situation id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Authorization token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer 1|y6VIcnEFVMpQvOHStKU0dcXD8CKEdJ0nQuW3anF6"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Situation name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Situation description",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Situation status",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/situations/{project_id}": {
            "get": {
                "tags": [
                    "Situations"
                ],
                "summary": "Get all situations of the project",
                "description": "Get all situations of the project.",
                "operationId": "53546cc216da09db9a3d0242f19b32ac",
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Authorization token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer 1|y6VIcnEFVMpQvOHStKU0dcXD8CKEdJ0nQuW3anF6"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Situation": {
                "title": "Situation",
                "description": "Situation model",
                "xml": {
                    "name": "Situation"
                }
            },
...
        }
    }
}